move the useGetCategories hook to the shared components folder in order to make it more accessible and reusable across components
