+++ src/shared/hooks/useGetCategories.js
